refactor(parsers): split instruction module and simplify helpers
Break the 617-line instruction/mod.rs into focused files:
- types.rs: ParsedInstruction, ParsedField, ParsedFieldValue types
- registry.rs: ParserRegistry with IDL loading logic
- mod.rs: InstructionParser trait, define_parser! macro, helpers

Also:
- Remove unused params from ParserRegistry::parse_cpi_event
- Simplify verbose account-name building in token2022_program.rs
- Remove unnecessary local wrapper in associated_token_program.rs
- Clean up verbose debug logging in parse_cpi_event
- Apply cargo fmt to output module files
